pyproject.toml
[project]
name = "game-plugins"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
